Anonymize IP addresses before sending them to Google Analytics
The last octet of IPv4 addresses and the last 80 bits of IPv6 addresses
are now zeroed out *before* they are sent to Google Analytics. This
shouldn't make any difference in actual analytics results, as the
"aip" flag was already set. This meant that the IP addresses were
anonymized the same way after they hit the analytics servers.

Closes #23.