Explicitly redirect to https: from http: URLs
The new version of the App Engine Flexible Environment doesn't support
secure handlers.

These changes also enable support for HTTP (non-HTTPS) traps. There is
currently no support for HTTP traps in the UI, though. The protocol for
the request (either "http" or "https") gets stored. HTTP trap URLs redirect
to their HTTPS counterparts as usual. Closes #26.

If the NODE_ENV is not "production" and APP_BASE_URL starts with "http:"
then the redirect will be skipped. This is to help localhost testing.